update name of python script from access_transit to access_subway_and_access_ADA - auto refactor all_accessors
